Permutate the Values of a Site-Pair Table Variable, Internal Function

A function which randomizes the values of the given variables within a site-pair table. This function is called from the gdm.varImp function and should not need to be called directly by the user.

permutateVarSitePair(spTab, siteVarTab, indexTab, vName)

Arguments

spTab
A given site-pair table.

siteVarTab
A given site x variable table.

indexTab
A table of index values for the given site-pair table.

vName
Vector of variable names in both the site-pair and site-by-variable tables.

Value

A site-pair table, with variable values randomized from the original site-pair table.

Note

This function is called from the gdm.varImp function and the user should not need to access directly.

See also